# parte2/openapi.yaml
openapi: 3.0.3
info:
  title: DW Parte 2 - Biblioteca API
  version: 1.0.0
servers:
  - url: http://localhost:3000

tags:
  - name: Authors
  - name: Publishers
  - name: Categories
  - name: Books

paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK

  /api/authors:
    get:
      tags: [Authors]
      summary: Listar autores
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Author" }
    post:
      tags: [Authors]
      summary: Criar autor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthorCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Author" }
        "400": { description: Bad Request }

  /api/authors/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Authors]
      summary: Obter autor por ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Author" }
        "400": { description: Invalid id }
        "404": { description: Not Found }
    put:
      tags: [Authors]
      summary: Atualizar autor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthorCreate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Author" }
        "400": { description: Bad Request }
        "404": { description: Not Found }
    delete:
      tags: [Authors]
      summary: Apagar autor
      responses:
        "204": { description: No Content }
        "400": { description: Invalid id }
        "404": { description: Not Found }

  /api/publishers:
    get:
      tags: [Publishers]
      summary: Listar editoras (opcional filtro por país)
      parameters:
        - name: country
          in: query
          required: false
          schema: { type: string, example: Portugal }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Publisher" }
    post:
      tags: [Publishers]
      summary: Criar editora
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublisherCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Publisher" }
        "400": { description: Bad Request }

  /api/publishers/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Publishers]
      summary: Obter editora por ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Publisher" }
        "400": { description: Invalid id }
        "404": { description: Not Found }
    put:
      tags: [Publishers]
      summary: Atualizar editora
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublisherCreate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Publisher" }
        "400": { description: Bad Request }
        "404": { description: Not Found }
    delete:
      tags: [Publishers]
      summary: Apagar editora
      responses:
        "204": { description: No Content }
        "400": { description: Invalid id }
        "404": { description: Not Found }

  /api/categories:
    get:
      tags: [Categories]
      summary: Listar categorias
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
    post:
      tags: [Categories]
      summary: Criar categoria
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "400": { description: Bad Request }
        "409": { description: Already exists }

  /api/categories/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Categories]
      summary: Obter categoria por ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "400": { description: Invalid id }
        "404": { description: Not Found }
    put:
      tags: [Categories]
      summary: Atualizar categoria
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryCreate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "400": { description: Bad Request }
        "404": { description: Not Found }
        "409": { description: Already exists }
    delete:
      tags: [Categories]
      summary: Apagar categoria
      responses:
        "204": { description: No Content }
        "400": { description: Invalid id }
        "404": { description: Not Found }

  /api/books:
    get:
      tags: [Books]
      summary: Listar livros (com filtros opcionais)
      parameters:
        - name: authorId
          in: query
          required: false
          schema: { type: integer, example: 1 }
        - name: publisherId
          in: query
          required: false
          schema: { type: integer, example: 1 }
        - name: title
          in: query
          required: false
          schema: { type: string, example: ensaio }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BookView" }
    post:
      tags: [Books]
      summary: Criar livro
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Book" }
        "400": { description: Bad Request }

  /api/books/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Books]
      summary: Obter livro por ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookView" }
        "400": { description: Invalid id }
        "404": { description: Not Found }
    put:
      tags: [Books]
      summary: Atualizar livro
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookCreate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Book" }
        "400": { description: Bad Request }
        "404": { description: Not Found }
    delete:
      tags: [Books]
      summary: Apagar livro
      responses:
        "204": { description: No Content }
        "400": { description: Invalid id }
        "404": { description: Not Found }

components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }

  schemas:
    Author:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        bio: { type: string, nullable: true }
        created_at: { type: string }
      required: [id, name]

    AuthorCreate:
      type: object
      properties:
        name: { type: string, example: Fernando Pessoa }
        bio: { type: string, nullable: true, example: Poeta }
      required: [name]

    Publisher:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        country: { type: string, nullable: true }
        created_at: { type: string }
      required: [id, name]

    PublisherCreate:
      type: object
      properties:
        name: { type: string, example: Porto Editora }
        country: { type: string, nullable: true, example: Portugal }
      required: [name]

    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
      required: [id, name]

    CategoryCreate:
      type: object
      properties:
        name: { type: string, example: Romance }
      required: [name]

    Book:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        isbn: { type: string, nullable: true }
        published_year: { type: integer, nullable: true }
        author_id: { type: integer }
        publisher_id: { type: integer }
        created_at: { type: string }
      required: [id, title, author_id, publisher_id]

    BookView:
      allOf:
        - $ref: "#/components/schemas/Book"
        - type: object
          properties:
            author_name: { type: string }
            publisher_name: { type: string }